Read below. Lots of changes, couldn't be bothered to make specific commits because of nested changes. * Verification now uses a generated id along with a number, which is supposed to be specific to the user that is connected to it. This uniqueness is not implemented yet, though that is the idea. * The way of getting a verification from the database is by supplying a verification number along with a user. This is simply because a verification cannot exist without a user, so we will pretty much never want to query for all verifications in the system. * Users have a list of verifications, and a verification belongs to one user. Pretty straightforward. * Changed the tests a bit, as well as their schema. * To save a verification, it is necessary to add it manually to the user (this is right now done in the BookkeepingController, but it should happen in some manager). Creating the verification and adding the user to the verification is not enough. This is because the user is the owner of the @OneToMany relationship, and as such only the user can cascade operations (cascade is actually not allowed in @ManyToOne tags).

+++ src/main/java/se/chalmers/bokforing/persistence/UserRepository.java
@@ -38,7 +38,7 @@ public interface UserRepository extends JpaRepository<UserAccount,Long> {
     
     @Modifying(clearAutomatically = true)
     @Query("update UserAccount u set u.email = ?1 where u.id = ?2")
-    int updateEmail(String newEmail, int id);
+    int updateEmail(String newEmail, Long id);
     
     @Modifying(clearAutomatically = true)
     @Query("update UserAccount u set u.userGroup = ?1 where u.email = ?2")
